feat: Add personal message box and reorder items
A new box for a personal message is added to the calendar. This
is not yet customizable, but the plan is to make it customizable
in the future.

The order of the items in the calendar is changed to ensure that
all corner parts can be covered. Previously, there were situations
where some corners could not be covered by any tile. This change
fixes that issue.

The file `jalali.scad` has been renamed to `data.scad` and now
contains both jalali and gregorian calendar data.